+++ src/app/auth.routes.ts
@@ -1,16 +1,20 @@
 import { Routes } from '@angular/router';
 
+import { loginOrRegisterGuard } from './auth/guards';
+
 export const authRoutes: Routes = [
   {
     path: 'login',
-    loadComponent: () => import('./auth/pages/login/login.component')
+    canMatch: [loginOrRegisterGuard],
+    loadComponent: () => import('./auth/pages/login/login.component'),
   },
   {
     path: 'register',
-    loadComponent: () => import('./auth/pages/register/register.component')
+    canMatch: [loginOrRegisterGuard],
+    loadComponent: () => import('./auth/pages/register/register.component'),
   },
   {
     path: '**',
-    redirectTo: 'login'
-  }
+    redirectTo: 'login',
+  },
 ];