+++ src/app/auth/interceptors/auth.interceptor.ts
@@ -1,0 +1,21 @@
+import { HttpHandlerFn, HttpRequest } from '@angular/common/http';
+import { inject } from '@angular/core';
+
+import { AuthService } from '../services';
+
+export const authInterceptor = (
+  req: HttpRequest<unknown>,
+  next: HttpHandlerFn
+) => {
+  const authService = inject(AuthService);
+  const token = authService.token();
+
+  if (token) {
+    const newReq = req.clone({
+      headers: req.headers.append('Authorization', `Bearer ${token}`),
+    });
+    return next(newReq);
+  }
+
+  return next(req);
+};