+++ src/app/auth/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './login.interface';
+export * from './register.interface';
+export * from './user-availability.interface';
+export * from './validate-token.interface';