+++ src/app/auth/interfaces/login.interface.ts
@@ -1,0 +1,18 @@
+export interface LoginRequest {
+  email: string;
+  password: string;
+}
+
+export interface LoginResponse {
+  user: UserLoginResponse;
+  token: string;
+}
+
+export interface UserLoginResponse {
+  _id: string;
+  email: string;
+  name: string;
+  surname: string;
+  createdAt: Date;
+  updatedAt: Date;
+}