+++ src/app/auth/interfaces/validate-token.interface.ts
@@ -1,0 +1,7 @@
+export interface ValidateTokenRequest {
+  token: string;
+}
+
+export interface ValidateTokenResponse {
+  isValidToken: boolean;
+}