+++ src/app/tasks/pages/task-details/task-details.component.ts
@@ -4,7 +4,6 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
   selector: 'app-task-details',
   imports: [],
   templateUrl: './task-details.component.html',
-  styleUrl: './task-details.component.css',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export default class TaskDetailsComponent { }
+export default class TaskDetailsComponent {}